+++ packages/firebase/src/model/notification/handlers/sendNotification.ts
@@ -1,10 +1,11 @@
 import { FastifyReply } from "fastify";
+import { MulticastMessage } from "firebase-admin/lib/messaging/messaging-api";
 import { SessionRequest } from "supertokens-node/framework/fastify";
 
 import { sendPushNotification } from "../../../lib";
 import DeviceService from "../../userDevice/service";
 
-import type { Message, TestNotificationInput } from "../../../types";
+import type { TestNotificationInput } from "../../../types";
 
 const testPushNotification = async (
   request: SessionRequest,
@@ -21,6 +22,7 @@ const testPushNotification = async (
   const {
     body,
     title,
+    data,
     userId: receiverId,
   } = request.body as TestNotificationInput;
 
@@ -46,9 +48,27 @@ const testPushNotification = async (
 
   const tokens = receiverDevices.map((device) => device.deviceToken as string);
 
-  const message: Message = {
+  const message: MulticastMessage = {
+    android: {
+      priority: "high",
+      notification: {
+        sound: "default",
+      },
+    },
+    apns: {
+      payload: {
+        aps: {
+          sound: "default",
+        },
+      },
+    },
     tokens,
     notification: {
+      title,
+      body,
+    },
+    data: {
+      ...data,
       title,
       body,
     },