+++ packages/multi-tenant/src/index.ts
@@ -25,11 +25,17 @@ export { default } from "./plugin";
 
 export { default as TenantService } from "./model/tenants/service";
 
+export * as validateTenantSchema from "./lib/validateTenantSchema";
+
 export { default as tenantMigrationPlugin } from "./migratePlugin";
 
 export { default as thirdPartyEmailPassword } from "./supertokens/recipes";
 
 export { default as invitationResolver } from "./invitations/resolver";
+
+export { default as tenantResolver } from "./model/tenants/resolver";
+
+export { default as tenantRoutes } from "./model/tenants/controller";
 
 export type {
   MultiTenantConfig,