+++ packages/user/src/model/invitations/handlers/createInvitation.ts
@@ -1,18 +1,15 @@
-import { getUsersByEmail } from "supertokens-node/recipe/thirdpartyemailpassword";
-
 import { ROLE_USER } from "../../../constants";
 import computeInvitationExpiresAt from "../../../lib/computeInvitationExpiresAt";
+import getInvitationService from "../../../lib/getInvitationService";
 import sendInvitation from "../../../lib/sendInvitation";
 import validateEmail from "../../../validator/email";
-import Service from "../service";
 
 import type {
   Invitation,
   InvitationCreateInput,
-  InvitationUpdateInput,
 } from "../../../types/invitation";
+import type { FilterInput } from "@dzangolab/fastify-slonik";
 import type { FastifyReply } from "fastify";
-import type { QueryResultRow } from "slonik";
 import type { SessionRequest } from "supertokens-node/framework/fastify";
 
 const createInvitation = async (
@@ -51,10 +48,18 @@ const createInvitation = async (
       });
     }
 
-    // check if user of the email already exists
-    const emailUser = await getUsersByEmail(email);
+    const service = getInvitationService(config, slonik, dbSchema);
 
-    if (emailUser.length > 0) {
+    const emailFilter = {
+      key: "email",
+      operator: "eq",
+      value: email,
+    } as FilterInput;
+
+    const userCount = await service.count(emailFilter);
+
+    // check if user of the email already exists
+    if (userCount > 0) {
       return reply.send({
         status: "ERROR",
         message: `User with email ${email} already exists`,
@@ -71,12 +76,12 @@ const createInvitation = async (
     const app = config.apps?.find((app) => app.id == appId);
 
     if (app) {
-      if (app.supportedRoles.includes(role)) {
+      if (app.supportedRoles.includes(invitationCreateInput.role)) {
         invitationCreateInput.appId = appId;
       } else {
         return reply.send({
           status: "ERROR",
-          message: `App ${app.name} does not support role ${role}`,
+          message: `App ${app.name} does not support role ${invitationCreateInput.role}`,
         });
       }
     }
@@ -84,12 +89,6 @@ const createInvitation = async (
     if (Object.keys(payload || {}).length > 0) {
       invitationCreateInput.payload = JSON.stringify(payload);
     }
-
-    const service = new Service<
-      Invitation & QueryResultRow,
-      InvitationCreateInput,
-      InvitationUpdateInput
-    >(config, slonik, dbSchema);
 
     let invitation: Invitation | undefined;