+++ packages/user/src/model/invitations/resolver.ts
@@ -1,28 +1,20 @@
 import { formatDate } from "@dzangolab/fastify-slonik";
 import mercurius from "mercurius";
 import { createNewSession } from "supertokens-node/recipe/session";
-import {
-  emailPasswordSignUp,
-  getUsersByEmail,
-} from "supertokens-node/recipe/thirdpartyemailpassword";
+import { emailPasswordSignUp } from "supertokens-node/recipe/thirdpartyemailpassword";
 
-import Service from "./service";
 import { ROLE_ADMIN } from "../../constants";
 import computeInvitationExpiresAt from "../../lib/computeInvitationExpiresAt";
+import getInvitationService from "../../lib/getInvitationService";
 import isInvitationValid from "../../lib/isInvitationValid";
 import sendInvitation from "../../lib/sendInvitation";
 import validateEmail from "../../validator/email";
 import validatePassword from "../../validator/password";
 
 import type { User } from "../../types";
-import type {
-  Invitation,
-  InvitationCreateInput,
-  InvitationUpdateInput,
-} from "../../types/invitation";
+import type { Invitation, InvitationCreateInput } from "../../types/invitation";
 import type { FilterInput, SortInput } from "@dzangolab/fastify-slonik";
 import type { MercuriusContext } from "mercurius";
-import type { QueryResultRow } from "slonik";
 
 const Mutation = {
   acceptInvitation: async (
@@ -65,11 +57,7 @@ const Mutation = {
         return mercuriusError;
       }
 
-      const service = new Service<
-        Invitation & QueryResultRow,
-        InvitationCreateInput,
-        InvitationUpdateInput
-      >(config, database, dbSchema);
+      const service = getInvitationService(config, database, dbSchema);
 
       const invitation = await service.findByToken(token);
 
@@ -164,10 +152,18 @@ const Mutation = {
         return mercuriusError;
       }
 
-      // check if user of the email already exists
-      const emailUser = await getUsersByEmail(email);
+      const service = getInvitationService(config, database, dbSchema);
 
-      if (emailUser.length > 0) {
+      const emailFilter = {
+        key: "email",
+        operator: "eq",
+        value: email,
+      } as FilterInput;
+
+      const userCount = await service.count(emailFilter);
+
+      // check if user of the email already exists
+      if (userCount > 0) {
         const mercuriusError = new mercurius.ErrorWithProps(
           `User with email ${email} already exists`
         );
@@ -185,11 +181,11 @@ const Mutation = {
       const app = config.apps?.find((app) => app.id == appId);
 
       if (app) {
-        if (app.supportedRoles.includes(role)) {
+        if (app.supportedRoles.includes(invitationCreateInput.role)) {
           invitationCreateInput.appId = appId;
         } else {
           const mercuriusError = new mercurius.ErrorWithProps(
-            `App ${app.name} does not support role ${role}`
+            `App ${app.name} does not support role ${invitationCreateInput.role}`
           );
 
           return mercuriusError;
@@ -199,12 +195,6 @@ const Mutation = {
       if (Object.keys(payload || {}).length > 0) {
         invitationCreateInput.payload = JSON.stringify(payload);
       }
-
-      const service = new Service<
-        Invitation & QueryResultRow,
-        InvitationCreateInput,
-        InvitationUpdateInput
-      >(config, database, dbSchema);
 
       let invitation: Invitation | undefined;
 
@@ -251,11 +241,7 @@ const Mutation = {
   ) => {
     const { app, config, database, dbSchema, reply } = context;
 
-    const service = new Service<
-      Invitation & QueryResultRow,
-      InvitationCreateInput,
-      InvitationUpdateInput
-    >(config, database, dbSchema);
+    const service = getInvitationService(config, database, dbSchema);
 
     const invitation = await service.findById(arguments_.id);
 
@@ -287,11 +273,11 @@ const Mutation = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service<
-      Invitation & QueryResultRow,
-      InvitationCreateInput,
-      InvitationUpdateInput
-    >(context.config, context.database, context.dbSchema);
+    const service = getInvitationService(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     let invitation = await service.findById(arguments_.id);
 
@@ -330,11 +316,11 @@ const Query = {
     context: MercuriusContext
   ) => {
     try {
-      const service = new Service<
-        Invitation & QueryResultRow,
-        InvitationCreateInput,
-        InvitationUpdateInput
-      >(context.config, context.database, context.dbSchema);
+      const service = getInvitationService(
+        context.config,
+        context.database,
+        context.dbSchema
+      );
 
       const invitation = await service.findByToken(arguments_.token);
 
@@ -361,11 +347,11 @@ const Query = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service<
-      Invitation & QueryResultRow,
-      InvitationCreateInput,
-      InvitationUpdateInput
-    >(context.config, context.database, context.dbSchema);
+    const service = getInvitationService(
+      context.config,
+      context.database,
+      context.dbSchema
+    );
 
     return await service.list(
       arguments_.limit,