+++ packages/user/src/model/roles/handlers/index.ts
@@ -1,0 +1,11 @@
+import createRole from "./createRole";
+import getPermissions from "./getPermissions";
+import getRoles from "./getRoles";
+import updatePermissions from "./updatePermissions";
+
+export default {
+  createRole,
+  getRoles,
+  getPermissions,
+  updatePermissions,
+};