+++ packages/user/src/model/users/resolver.ts
@@ -4,8 +4,8 @@ import { emailPasswordSignUp } from "supertokens-node/recipe/thirdpartyemailpass
 import UserRoles from "supertokens-node/recipe/userroles";
 
 import filterUserUpdateInput from "./filterUserUpdateInput";
-import Service from "./service";
 import { ROLE_ADMIN } from "../../constants";
+import getUserService from "../../lib/getUserService";
 import validateEmail from "../../validator/email";
 import validatePassword from "../../validator/password";
 
@@ -70,7 +70,13 @@ const Mutation = {
 
       // signup
       const signUpResponse = await emailPasswordSignUp(email, password, {
+        autoVerifyEmail: true,
         roles: [ROLE_ADMIN],
+        _default: {
+          request: {
+            request: reply.request,
+          },
+        },
       });
 
       if (signUpResponse.status !== "OK") {
@@ -123,11 +129,7 @@ const Mutation = {
       return mercuriusError;
     }
 
-    if (context.roles === undefined || !context.roles.includes(ROLE_ADMIN)) {
-      return new mercurius.ErrorWithProps(`User is not an admin`, {}, 403);
-    }
-
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -150,11 +152,7 @@ const Mutation = {
   ) => {
     const { id } = arguments_;
 
-    if (context.roles === undefined || !context.roles.includes(ROLE_ADMIN)) {
-      return new mercurius.ErrorWithProps(`User is not an admin`, {}, 403);
-    }
-
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -176,7 +174,7 @@ const Mutation = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -214,7 +212,7 @@ const Mutation = {
   ) => {
     const { data } = arguments_;
 
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -283,7 +281,7 @@ const Query = {
     arguments_: Record<string, never>,
     context: MercuriusContext
   ) => {
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -310,7 +308,7 @@ const Query = {
     arguments_: { id: string },
     context: MercuriusContext
   ) => {
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema
@@ -328,7 +326,7 @@ const Query = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service(
+    const service = getUserService(
       context.config,
       context.database,
       context.dbSchema